requirements.txt
python-dotenv==0.17.1
tweepy==4.10.0
kafka-python==2.0.2
vaderSentiment==3.3.2
hdfs==2.7.0
pyspark==3.2.1
streamlit==1.9.2
matplotlib==3.5.2
matplotlib-inline==0.1.3
seaborn==0.11.2